refactor: move subWindows to correct folder
