+++ Source/ChironEngine/ChironEngine.vcxproj
@@ -45,8 +45,8 @@
     <ClCompile Include="Source\DataModels\Window\EditorWindow\ConsoleWindow.cpp" />
     <ClCompile Include="Source\DataModels\Window\EditorWindow\EditorWindow.cpp" />
     <ClCompile Include="Source\DataModels\Window\EditorWindow\SceneWindow.cpp" />
-    <ClCompile Include="Source\DataModels\Window\EditorWindow\SubWindows\CameraSubWindow.cpp" />
-    <ClCompile Include="Source\DataModels\Window\EditorWindow\SubWindows\EngineStatisticsSubWindow.cpp" />
+    <ClCompile Include="Source\DataModels\Window\EditorWindow\SubWindows\Configuration\CameraSubWindow.cpp" />
+    <ClCompile Include="Source\DataModels\Window\EditorWindow\SubWindows\Configuration\EngineStatisticsSubWindow.cpp" />
     <ClCompile Include="Source\DataModels\Window\EditorWindow\SubWindows\SubWindow.cpp" />
     <ClCompile Include="Source\DataModels\Window\Window.cpp" />
     <ClCompile Include="Source\DataModels\Window\MainMenuWindow.cpp" />
@@ -75,8 +75,8 @@
     <ClInclude Include="Source\DataModels\Window\EditorWindow\ConsoleWindow.h" />
     <ClInclude Include="Source\DataModels\Window\EditorWindow\EditorWindow.h" />
     <ClInclude Include="Source\DataModels\Window\EditorWindow\SceneWindow.h" />
-    <ClInclude Include="Source\DataModels\Window\EditorWindow\SubWindows\CameraSubWindow.h" />
-    <ClInclude Include="Source\DataModels\Window\EditorWindow\SubWindows\EngineStatisticsSubWindow.h" />
+    <ClInclude Include="Source\DataModels\Window\EditorWindow\SubWindows\Configuration\CameraSubWindow.h" />
+    <ClInclude Include="Source\DataModels\Window\EditorWindow\SubWindows\Configuration\EngineStatisticsSubWindow.h" />
     <ClInclude Include="Source\DataModels\Window\EditorWindow\SubWindows\SubWindow.h" />
     <ClInclude Include="Source\DataModels\Window\Window.h" />
     <ClInclude Include="Source\DataModels\Window\MainMenuWindow.h" />
