+++ Source/ChironEngine/ChironEngine.vcxproj.filters
@@ -228,10 +228,10 @@
     <ClCompile Include="Source\DataModels\Window\EditorWindow\SubWindows\SubWindow.cpp">
       <Filter>Source\DataModels\Windows\EditorWindow\SubWindows</Filter>
     </ClCompile>
-    <ClCompile Include="Source\DataModels\Window\EditorWindow\SubWindows\EngineStatisticsSubWindow.cpp">
+    <ClCompile Include="Source\DataModels\Window\EditorWindow\SubWindows\Configuration\EngineStatisticsSubWindow.cpp">
       <Filter>Source\DataModels\Windows\EditorWindow\SubWindows\Configuration</Filter>
     </ClCompile>
-    <ClCompile Include="Source\DataModels\Window\EditorWindow\SubWindows\CameraSubWindow.cpp">
+    <ClCompile Include="Source\DataModels\Window\EditorWindow\SubWindows\Configuration\CameraSubWindow.cpp">
       <Filter>Source\DataModels\Windows\EditorWindow\SubWindows\Configuration</Filter>
     </ClCompile>
   </ItemGroup>
@@ -414,10 +414,10 @@
     <ClInclude Include="Source\DataModels\Window\EditorWindow\SubWindows\SubWindow.h">
       <Filter>Source\DataModels\Windows\EditorWindow\SubWindows</Filter>
     </ClInclude>
-    <ClInclude Include="Source\DataModels\Window\EditorWindow\SubWindows\EngineStatisticsSubWindow.h">
+    <ClInclude Include="Source\DataModels\Window\EditorWindow\SubWindows\Configuration\EngineStatisticsSubWindow.h">
       <Filter>Source\DataModels\Windows\EditorWindow\SubWindows\Configuration</Filter>
     </ClInclude>
-    <ClInclude Include="Source\DataModels\Window\EditorWindow\SubWindows\CameraSubWindow.h">
+    <ClInclude Include="Source\DataModels\Window\EditorWindow\SubWindows\Configuration\CameraSubWindow.h">
       <Filter>Source\DataModels\Windows\EditorWindow\SubWindows\Configuration</Filter>
     </ClInclude>
   </ItemGroup>
