+++ Source/ChironEngine/Source/DataModels/Window/EditorWindow/SubWindows/Configuration/CameraSubWindow.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "SubWindow.h"
+#include "../SubWindow.h"
 
 class CameraSubWindow : public SubWindow
 {
